more button work, added releaseHook commands

+++ 2018 Robot Code/src/org/usfirst/frc/team5822/robot/commands/ArmButton.java
@@ -7,7 +7,7 @@ import edu.wpi.first.wpilibj.command.CommandGroup;
  */
 public class ArmButton extends CommandGroup {
 
-    public ArmButton(String area, int degrees) 
+    public ArmButton(String area, double degrees) 
     {
        if (area == "Floor")
        {
